feat: added sponsor menu and sponsor overlay + refactored bitcoin/bond/stake screens
