--- src-vue/app-treasury/interfaces/IPortfolioTab.ts
@@ -1,6 +1,0 @@
-export enum PortfolioTab {
-  Overview = 'Overview',
-  AssetBreakdown = 'AssetBreakdown',
-  ProfitAnalysis = 'ProfitAnalysis',
-  TransactionHistory = 'TransactionHistory',
-}
