--- src-vue/app-treasury/panels/portfolio/ProfitAnalysis.vue
@@ -1,6 +1,0 @@
-<template>
-  <div class="flex h-full flex-row items-center justify-center pb-[20%] text-center">
-    <div class="text-6xl font-bold text-slate-600/30">Coming Soon</div>
-  </div>
-</template>
-<script setup lang="ts"></script>
