+++ src-vue/screens/mining-screen/Dashboard.vue
@@ -66,7 +66,7 @@
             <span v-if="miningReturnSummary.percent !== undefined">
               {{ numeral(miningReturnSummary.percent).formatIfElseCapped('< 100', '0.[00]', '0,0', 9_999) }}%
             </span>
-            <span v-else>--</span>
+            <span>--</span>
             <label>Mining RTD</label>
           </TooltipTrigger>
           <TooltipContent side="bottom" :sideOffset="-10" align="end" :collisionPadding="9" class="text-right bg-white border border-gray-800/20 rounded-md shadow-2xl z-50 p-4 w-xs text-slate-900/60">
@@ -142,12 +142,12 @@
                                 : `${seconds} Second${seconds === 1 ? '' : 's'}`
                           }}
                         </template>
-                        <template v-else>
+                        <template>
                           Your Next Bid Pending
                         </template>
                       </div>
                     </CountdownClock>
-                    <div v-else>{{ nextBidTimingLabel }}</div>
+                    <div>{{ nextBidTimingLabel }}</div>
                     <div>{{ auctionStatsLabel }}</div>
                     <div>{{ formatBidAmount(lowestWinningBid) }} Is the Lowest Bid</div>
                     <div>{{ formatBidAmount(myLastBidMicrogons) }} Was Your Last Bid</div>
@@ -166,12 +166,12 @@
                                 : `${seconds} Second${seconds === 1 ? '' : 's'}`
                           }}
                         </template>
-                        <template v-else>
+                        <template>
                           Auction May Close Any Moment
                         </template>
                       </div>
                     </CountdownClock>
-                    <div v-else class="titleize">{{ auctionTimingLabel }}</div>
+                    <div class="titleize">{{ auctionTimingLabel }}</div>
                   </div>
                 </div>
               </div>
